Eth blocks - first load sequence in progress - use state for storing ethBlocks
