+++ pages/playground/eth-blocks/index.vue
@@ -2,16 +2,17 @@
   <div class="eth-blocks-page page-container eth-base-text">
     <div
       id="ethBlocks"
-      ref="ethBlocks"
+      ref="ethBlocksWrapper"
       class="eth-blocks"
       :style="{
         paddingTop: blocksBasePosition + 'px',
       }"
     >
+      <!--      :ref="(el) => addNewBlockEl(el, block.blockId)"-->
+
       <div
         v-for="block in blocksToRender"
         :key="block.blockId"
-        :ref="(el) => addNewBlockEl(el, block.blockId)"
         v-action-on-scroll="{
           activeRange: 1,
           bidirectionalActivation: true,
@@ -36,8 +37,6 @@ import {
   generateBlockData,
   deserializeBlock,
   generateLoadingBlockData,
-  enterAni,
-  blockContentAniIn,
 } from "~/utils/playground/eth-blocks/web3-helpers";
 import { gsap } from "gsap";
 import SplitText from "gsap/SplitText";
@@ -47,6 +46,11 @@ import type {
   BlockExtended,
   BlockItem,
 } from "#shared/types/playground/eth-blocks";
+import {
+  blockContentAniIn,
+  enterAni,
+} from "~/utils/playground/eth-blocks/eth-block-animation-helpers";
+import { useEthBlocks } from "~/stores/playground/eth-blocks-store";
 gsap.registerPlugin(SplitText);
 
 //**************************
@@ -55,18 +59,18 @@ gsap.registerPlugin(SplitText);
 
 const maxBlocks = 50;
 
-const ethBlocks = ref<HTMLElement | null>(null);
+const ethBlocksWrapper = ref<HTMLElement | null>(null);
 const blocksBasePosition = ref(0);
-const blocks = ref<Map<string, BlockItem>>(new Map());
+const { ethBlocks } = useEthBlocks();
 const blocksToRender = computed<BlockItem[]>(() => {
-  return [...blocks.value.values()].sort((a, b) =>
+  return [...ethBlocks.value.values()].sort((a, b) =>
     a.blockId > b.blockId ? -1 : 1,
   );
 });
 
 const tlNewBlockAniIn = gsap.timeline({
-  paused: true,
-  onComplete: () => {},
+  // paused: true,
+  // onComplete: () => {},
 });
 
 let eventSource: EventSource;
@@ -79,18 +83,17 @@ const fetchInitialBlocks = async () => {
   const { data: initialBlocks } = await useFetch(
     "/api/playground/eth-blocks/latest",
   );
-  initialBlocks.value?.forEach((raw: BlockExtended, index: number) => {
+  initialBlocks.value?.forEach((raw: BlockExtended) => {
     const block = deserializeBlock(raw);
-    const blockId =
-      index === 0
-        ? "block_0"
-        : new Date().getTime().toString() + `_latest_${index}`;
-    const loadingBlock = generateLoadingBlockData(blockId);
-    blocks.value.set(blockId, generateBlockData(block, loadingBlock));
+    const blockId = blockIdCounter.value;
+    const loadingBlock = generateLoadingBlockData(blockId.toString());
+    ethBlocks.value.set(
+      blockId.toString(),
+      generateBlockData(block, loadingBlock),
+    );
+    blockIdCounter.value += 1;
   });
 };
-
-// const firstLoaderAni = async () => {};
 
 const blockIdCounter = ref(0);
 const newBlockId = computed(() => {
@@ -100,9 +103,9 @@ const newBlockId = computed(() => {
 async function newLoadingBlock(firstAnimation = false) {
   ethBlocksAnimation.loadingBlockId = newBlockId.value;
   const blockData = generateLoadingBlockData(newBlockId.value);
-  blocks.value.set(ethBlocksAnimation.loadingBlockId, blockData);
+  ethBlocks.value.set(ethBlocksAnimation.loadingBlockId, blockData);
   await nextTick();
-  const el = blocks.value.get(newBlockId.value)?.elRef as HTMLElement;
+  const el = getElFromBlockId(newBlockId.value);
   if (!el) {
     return;
   }
@@ -120,7 +123,7 @@ async function newLoadingBlock(firstAnimation = false) {
     const blockProgressBarEl = el.querySelector(".block-loading-progress");
     tlNewBlockAniIn.to(blockProgressBarEl, {
       width: "100%",
-      duration: firstAnimation ? 1 : ethBlocksAnimation.blockLoadingTime,
+      duration: firstAnimation ? 5 : ethBlocksAnimation.blockLoadingTime,
       onComplete: () => {
         resolve(true);
       },
@@ -128,55 +131,63 @@ async function newLoadingBlock(firstAnimation = false) {
   });
 }
 
-function addNewBlockEl(
-  el: Element | ComponentPublicInstance | null,
-  blockId: string,
-) {
-  const block = blocks.value.get(blockId.toString());
-  if (block && !block.elRef) {
-    block.elRef = el;
-  }
-}
+// function addNewBlockEl(
+//   el: Element | ComponentPublicInstance | null,
+//   blockId: string,
+// ) {
+//   const block = blocks.value.get(blockId.toString());
+//   if (block && !block.elRef) {
+//     block.elRef = el;
+//   }
+// }
+
+const getElFromBlockId = (blockId: string) => {
+  const block = ethBlocks.value.get(blockId);
+  if (!block) return;
+  return block.elRef as HTMLElement;
+};
+
+// const progressBarFinish = () => {
+//   tlNewBlockAniIn.tweenTo(tlNewBlockAniIn.duration(), {
+//     duration: 0.3,
+//     ease: "linear",
+//   });
+// };
+
+const resetProgressBar = (blockId: string) => {
+  const el = getElFromBlockId(blockId);
+  if (!el) return;
+  const blockProgressBarEl = el.querySelector(".block-loading-progress");
+  if (!blockProgressBarEl) return;
+  tlNewBlockAniIn.to(blockProgressBarEl, {
+    width: 0,
+    duration: 0,
+    opacity: 1,
+  });
+};
 
 const blockDoneAnimate = async (blockId: string) => {
-  const block = blocks.value.get(blockId);
-  if (!block) return;
-  const el = block.elRef as HTMLElement;
+  const el = getElFromBlockId(blockId);
   if (!el) return;
 
   return new Promise((resolve) => {
-    const addTimelineAnimations = () => {
-      tlNewBlockAniIn.to(el.querySelector(".block-loading-progress"), {
-        width: "0%",
-        duration: 0.15,
-        right: 0,
-        left: "initial",
-      });
-
+    const addTimelineAnimations = async () => {
       tlNewBlockAniIn.fromTo(
         el,
         { height: "10px" },
         { height: "236px", duration: 0.5, marginTop: "20px" },
       );
-
       blockContentAniIn(el, tlNewBlockAniIn);
-
-      tlNewBlockAniIn.to(el.querySelector(".block-loading-progress"), {
-        width: 0,
-        duration: 0,
-        opacity: 1,
-        onComplete: () => {
-          resolve(true);
-        },
+      tlNewBlockAniIn.call(() => {
+        resolve(true);
       });
-
-      // tlNewBlockAniIn.play();
     };
 
     tlNewBlockAniIn.tweenTo(tlNewBlockAniIn.duration(), {
       duration: 0.3,
       ease: "linear",
       onComplete: () => {
+        resetProgressBar(blockId);
         addTimelineAnimations();
       },
     });
@@ -187,37 +198,36 @@ const addBlockListener = () => {
   eventSource = new EventSource("/api/playground/eth-blocks/watch");
   eventSource.onmessage = async ({ data }) => {
     const block = deserializeBlock(JSON.parse(data));
-    const loadingBlock = blocks.value.get(ethBlocksAnimation.loadingBlockId);
+    const loadingBlock = ethBlocks.value.get(ethBlocksAnimation.loadingBlockId);
     if (!loadingBlock) return;
-    blocks.value.set(
+    ethBlocks.value.set(
       ethBlocksAnimation.loadingBlockId,
       generateBlockData(block, loadingBlock),
     );
     await nextTick();
     await blockDoneAnimate(ethBlocksAnimation.loadingBlockId);
     await newLoadingBlock();
-    if (blocks.value.size > maxBlocks) {
-      const oldestKey = blocks.value.keys().next().value;
-      if (oldestKey) blocks.value.delete(oldestKey);
+    if (ethBlocks.value.size > maxBlocks) {
+      const oldestKey = ethBlocks.value.keys().next().value;
+      if (oldestKey) ethBlocks.value.delete(oldestKey);
     }
   };
 };
 
 onUnmounted(() => eventSource?.close());
 
+fetchInitialBlocks();
 onMounted(async () => {
+  console.log("mounted");
   ethBlocksAnimation.setBlockBasePosition();
   blocksBasePosition.value = ethBlocksAnimation.blocksBasePosition;
-  if (ethBlocks.value) {
-    await ethBlocksAnimation.init(ethBlocks.value);
+  if (ethBlocksWrapper.value) {
+    await ethBlocksAnimation.init(ethBlocksWrapper.value);
+    ethBlocksAnimation.loadTextures(); // all final textures
   }
-  await newLoadingBlock(true);
-
-  await fetchInitialBlocks();
-
-  await enterAni(tlNewBlockAniIn);
+  enterAni(tlNewBlockAniIn, ethBlocksWrapper.value);
   addBlockListener();
-  await ethBlocksAnimation.loadTextures(); // all final textures
+  await newLoadingBlock();
 });
 
 // https://www.shadertoy.com/view/wccSDf
@@ -247,6 +257,7 @@ onMounted(async () => {
   margin: 0 auto;
   height: 0;
   width: 0;
+  opacity: 0;
   border-radius: 25px;
   &:not(.block-loading) {
     opacity: 0;
